Add sys_brk (#139)
[brk](https://www.man7.org/linux/man-pages/man2/brk.2.html) change the
location of the program break, which defines the end of the process's
data segment (i.e., the program break is the first location after the
end of the uninitialized data segment). Increasing the program break has
the effect of allocating memory to the process; decreasing the break
deallocates memory.

A typical Linux process memory layout:
```
+--------------------------+ 0xC0000000 or 0x7ffffffff000
|     Kernel Space         | (not accessible from userland)
+--------------------------+
|     Stack (grows down)   |
+--------------------------+
|     Mmap area (down)     |
|--------------------------| ← mmap base (e.g., 0x7fff0000)
|     Shared libraries     |
|     Heap (brk, grows up) |
|--------------------------| ← brk start
|     BSS / Data / Text    |
|--------------------------| ← program load addr (e.g., 0x400000)
```

To avoid memory overlapping, the `brk` heap starts after the end of the
binary's .data section and goes upward, while the mmap allocations
typically start just below the stack and go downward. This PR also fixes
`get_unmmaped_area` so that it returns memory from high addresses to low
addresses, conforming to the Linux kernel's memory layout.

Also add a new API `reserved_pages` to make certain regions as
allocated.