Revert "check for changed .rs files before semver checks" (#168)
This reverts commit fcb17a1308ee245c0862be530cab0da3185f3975 (part of
#165).

The change (intended at speeding things up) is over-eager, and makes the
entire semver-checks workflow ineffective. By reverting it out, we make
sure that we are actually using the more effective version. At some
point in the future, I might try to bring back some version of this to
speed things up.