Multithreading-friendly mock platform (#238)
This PR finishes up some more implementation in our mock platform to
make it multi-thread friendly. Basically, it replaces the usage of
`RefCell` with `RwLock` and finishes up an implementation of `RawMutex`.

Context: our Mock platform left parts of it unimplemented because none
of our tests needed it. However, and upcoming change I'll be pushing in
a new PR will need to migrate some tests from the shim to the litebox
crate, and then those tests cannot depend on a real platform anymore,
and thus they need a bit more support from our mock platform.