Implement `read_dir` for the file systems (#247)
This PR introduces `read_dir` to all the file systems. The layered file
system is the most interesting one in terms of implementation, because
you need to account for directories with files existing across the
layers.

Along the way, this PR also fixes up a minor issue that existed in the
`tar-ro` file system where trying to open the root directory would not
work (I discovered it when writing tests for `read_dir`).