Simplify test code (#300)
This PR refactors the testing code so that it is easier to add a new
end-to-end test.

1. Use `ldd` to find all dependencies instead of manually specifying
them.
2. Add a new test `test_static_exec_with_rewriter` for syscall rewriter
backend (which can run on both 64bit and 32bit platform as it does not
rely on `rtld_audit`).
3. Add a new end-to-end test `thread.c`. For a new test, we just need to
put its C code into the test folder.
4. I also re-enabled `test_runner_with_nodejs` and it seems to work fine
now (I manually triggered CI multiple times). I haven't figure out the
root cause of it #240 as I cannot reproduce it locally. #309 may have
fixed it, but I'm not sure.
5. Add a test `test_node_with_rewriter` for node with rewriter backend.
It's disabled for now as it takes > 5min to rewrite the binary.
Alternatively, we could upload rewritten binaries to avoid it.
6. Seccompiler does not support x86 arch yet; only enable those tests requiring seccomp on x86_64.