Migrate pipes to litebox crate (#333)
This PR migrates the pipes support from the Linux shim over to the
LiteBox crate. Along the way, it fixes a bug that existed in the older
pipes implementation, and also improves the documentation for the pipes.

Interestingly, I came up with a nicer design for pipes (see
7d4bcbb3a0612841ab7ac833469844b995854878, in branch
`prototyping/jayb/nicer-pipes-interface-but-rustdoc-bug-139964`), but
unfortunately, it hits a rustdoc bug
(https://github.com/rust-lang/rust/issues/139964) where it compiles and
works fine, but sends `cargo doc` into an infinite loop. Thus sadly, we
are not going to be able to use the nicer design until that rustdoc bug
is fixed. Technically, the next generation trait solver (`
-Znext-solver`) for rustdoc fixes this issue, but that is not yet
stabilizied (that's why it has the `-Z`) in it. I'm writing all this
down to document that we are _currently_ not using the nicer design, but
in the future, once that bug is fixed, we should be able to move to the
nicer design.