Fix futex manager (#335)
Not sure if it is the same bug I found in #330. I can reproduce a race
condition issue by running `test_node_with_seccomp` multiple times ( >
50).
 
The root cause was that a waker may operate on different `lockable`
objects while it is waiting for waiters to wake up. This PR addresses it
by assigning a unique number to a `lockable` so that the waker can know
whether it gets the same `lockable` object or not.