Add component detection manifest + NOTICE for `sync::Mutex` and `sync::RwLock` (#567)
Even though our implementations have diverged a significant amount, the
original code was from the Rust source. In preparation for open-sourcing
LiteBox, the `cgmanifest.json` helps us document the source of these in
a machine-readable form. Additionally, `NOTICE.txt` covers the notice
for this at the root of the dir.