Automated boilerplate copyright header checks (#569)
This PR adds the necessary copyright headers to all our source files,
and introduces an automated CI test for it, so we don't forget them. See
`dev_tests/src/boilerplate.rs` for the core implementation; most of the
rest of this PR is the actual introduction of the boilerplate.

This resolves #561